apps/server/pyproject.toml
[project]
name = "chatty-server"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "sse-starlette>=2.1.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.5.0",
    "pydantic[email]>=2.5.0",
    "bcrypt>=4.2.0",
    "python-jose[cryptography]>=3.3.0",
    "psycopg[binary]>=3.2.0",
    "psycopg-pool>=3.2.0",
    "httpx>=0.27.0",
    "redis>=7.4.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "basedpyright>=1.18.0",
    "ruff>=0.8.0",
    "anyio>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812", "ISC001",
    "D100", "D101", "D102", "D103", "D104", "D107",
    "D203", "D212",
    "FIX002", "TD002", "TD003",
    "UP007", "TC001", "TC002", "TC003",
    "B008",
    "S105", "S107",
    "FAST002",
    "ARG001", "ARG002",
    "INP001",
    "FBT001", "FBT002",
    "TRY003",
    "EM101",
    "RUF005", "RUF006",
    "PLR0911",
    "PLC0415",
    "PERF401",
    "SIM105", "SIM110",
    "RET504",
    "ERA001",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "S101", "ANN", "PLR2004", "SLF001", "ARG001", "ARG002",
    "S106", "S107", "S108",
    "E501",
    "D",
    "B",
    "PT",
]
"migrations/**" = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = [
    "fastapi.Depends", "fastapi.Query", "fastapi.Body",
    "fastapi.Header", "fastapi.Path", "fastapi.Form",
    "fastapi.File", "fastapi.Cookie", "fastapi.Security",
    "Depends", "Query", "Body", "Header", "Path", "Form", "File", "Cookie", "Security",
]

[tool.ruff.lint.pylint]
max-args = 10
max-returns = 4
max-branches = 10

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.basedpyright]
pythonVersion = "3.12"
venvPath = "../.."
venv = ".venv"
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnknownVariableType = true
reportUnknownParameterType = true
reportUnknownMemberType = false
reportMissingParameterType = true
reportUntypedFunctionDecorator = false
reportPrivateUsage = true
reportUninitializedInstanceVariable = true
reportCallInDefaultInitializer = true
reportImplicitOverride = true
